requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
opencv-python>=4.5.0

